services/fraud-detection-service/requirements.txt
flask==3.0.0
psycopg2-binary==2.9.9
kafka-python==2.0.2
prometheus-client==0.19.0
flasgger==0.9.7.1
